backend/requirements.txt
# ============================================================================
# Binance Portfolio Manager - Backend Requirements
# ============================================================================

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.35  # Compatible avec Python 3.13

# Binance API
python-binance==1.0.19
cryptography==41.0.7
requests==2.31.0
websocket-client==1.6.4

# Environment & Utils
python-dotenv==1.0.0

# Date/Time utilities
python-dateutil==2.8.2
